splitting out marc functions
